requirements.txt
cmake==3.25.2
dlib==19.24.0
face-recognition==1.3.0
numpy==1.24.2
Pillow==11.0.0
playsound==1.3.0
gtts==2.5.4
vosk=0.3.45
sounddevice==0.5.1
soundfile==0.12.1
mediapipe==0.10.20
google-api-python-client==2.157.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
PyQt5==5.15.11
schedule==1.2.2